+++ bindings/muen/muen-clock.c
@@ -87,8 +87,8 @@ int tscclock_init(uint64_t freq __attribute__((unused)))
     tsc_mult = (NSEC_PER_SEC << 32) / tsc_freq;
     min_delta = (tsc_freq + (NSEC_PER_SEC - 1)) / NSEC_PER_SEC;
     time_base = 0;
-    log(INFO, "Solo5: Clock source: Muen PV clock, TSC frequency %lu Hz\n",
-        tsc_freq);
+    log(INFO, "Solo5: Clock source: Muen PV clock, TSC frequency %llu Hz\n",
+        (unsigned long long)tsc_freq);
     return 0;
 }
